backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.20.0
python-multipart>=0.0.6
opencv-python>=4.8.0
numpy>=1.24.0
Pillow>=10.0.0
pydantic>=2.0.0
python-dotenv>=1.0.0
aiofiles>=23.0.0
insightface>=0.7.3
onnxruntime>=1.15.0
